Change docs page layout from flex to grid.
Note that this required introducing some additional js to set the number of content sections (grid rows) and also toa dd classes for the first and last sections.

first and last section classes are required to set the proper block padding due to the way things are ordered.

The number of grid rows is required so the sidebars (toc and mini-toc) can properly span all rows. It turns out setting from `grid-row: 1 / -1` only works when there is a defined number of rows. lacking that, there is no grid line numbered -1, and hence it doesn't work.
So we set a dynamic css property to the number of grid rows and use that to declare `grid-template-rows`.